+++ app/src/main/java/com/delta/smt/entity/StorageReady.java
@@ -21,7 +21,7 @@ public class StorageReady extends TimeEntity {
     private String work_order;
     private String side;
     private int status;
-    private double remain_time= 0L;
+    private double remain_time = 0L;
 
     public String getLine_name() {
         return line_name;
@@ -59,7 +59,7 @@ public class StorageReady extends TimeEntity {
         return remain_time;
     }
 
-    public void setRemain_time(Long remain_time) {
+    public void setRemain_time(double remain_time) {
         this.remain_time = remain_time;
     }
 }
